線形補間のコードを別ファイルに分けた
